Move elevation noise files into their own folder [Re: nmkataoka/adv-life#232]
Since there will be at least one more elevation noise generator coming,
I'm creating a folder for elevation noise generators.
